corrected customer model
