+++ Models/Customers.cs
@@ -1,0 +1,23 @@
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+ 
+namespace BangazonAPI.Models
+{
+   public class Customers
+   {
+     [Key]
+     public int CustomerId { get; set; }
+ 
+     [Required]
+     [DataType(DataType.Date)]
+     [DatabaseGenerated(DatabaseGeneratedOption.Computed)]
+     public DateTime DateCreated { get; set; }
+ 
+     [Required]
+     public string FirstName { get; set; }
+     public string LastName { get; set; }
+ 
+   }
+}
